+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New devices:
 - Bitron Wireless smart plug
 - ROBB ZigBee AC phase-cut dimmer
 - Eurotronic SPZB0001 thermostat
+- Gira ZigBee Light Link wall transmitter
 
 ## 0.0.19
 - Add temperature sensor support for Xiaomi Mi power plug